+++ src/Stryker.Core/Stryker.Core/Mutants/MutationContext.cs
@@ -8,13 +8,13 @@ namespace Stryker.Core.Mutants
     /// </summary>
     public class MutationContext: IDisposable
     {
-        private readonly MutantOrchestrator _mainOrchestrator;
+        private readonly CsharpMutantOrchestrator _mainOrchestrator;
         private readonly MutationContext _ancestor;
         public readonly List<Mutant> ExpressionLevelMutations = new List<Mutant>();
         public readonly List<Mutant> BlockLevelControlledMutations = new List<Mutant>();
         public readonly List<Mutant> StatementLevelControlledMutations = new List<Mutant>();
 
-        public MutationContext(MutantOrchestrator mutantOrchestrator)
+        public MutationContext(CsharpMutantOrchestrator mutantOrchestrator)
         {
             _mainOrchestrator = mutantOrchestrator;
         }
